Fix setup.m for matlabpath on parallel workers
The client path is shared with the workers at the time the parpool is started. However,
Fieldtrip does some additional path setting when it is first used, in addition to the
main folder needing to be on the path. Calling ft_defaults before starting the parpool
makes sure the paths are all set.