Add function and call it, to gather reference data for better timing info
Calls to a new save_input function are added to the bench and tutorial functions,
that save the input data to the analysis functions we want to time.
This addition is only for the instrumentation within this branch, and not meant
as a permanent addition to the bench and tutorial functions.

Timings_save_input is added as the main function to call to trigger the data
gathering.